func: add tela de vizualização de usuarios
